added templates for very basic extensions
One of the development goals of the Commotion_client (commotion-computer to the savvy) is to make extension of the toolkit as easy as possible for other developers. As such, this branch will focus on building a set of templates (shown in this commit) and extending them to create a full commotion core module.